pretty prints tidy table1 data frame returned from tidytable1

pp_table1 pretty print the tidy table1 data frame produced by tidytable1 by removing duplicate information.

pp_table1(table1_df, pp_vars = c("#missing"))

Arguments

table1_df
A data_frame returned from tidytable1.

pp_vars
A vector of table1 columns to be pretty print in addition to the 'Name' column.

Value

A data frame with duplicate information removed and ready for print out (e.g. pass it to pander::pander(ret, missing="")).